@@ -1,20 +1,30 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {DialogerComponent} from './dialoger.component';
-import {DialogComponent} from './dialog/dialog.component';
 import {MatButtonModule} from '@angular/material/button';
+import {MatToolbarModule} from '@angular/material/toolbar';
+import {MatIconModule} from '@angular/material/icon';
+import {PromiseButtonModule} from '@core/components';
+import {DialogWindowComponent} from './dialog-window/dialog-window.component';
+import {BasedialogcompComponent} from '@core/components/dialoger/inheritance/basedialogcomp/basedialogcomp.component';
 
 @NgModule({
   declarations: [
     DialogerComponent,
-    DialogComponent
+    DialogWindowComponent,
+    BasedialogcompComponent
   ],
   exports: [
-    DialogerComponent
+    DialogerComponent,
+    DialogWindowComponent,
+    BasedialogcompComponent
   ],
   imports: [
     CommonModule,
-    MatButtonModule
+    MatButtonModule,
+    MatToolbarModule,
+    MatIconModule,
+    PromiseButtonModule
   ]
 })
 export class DialogerModule {
