@@ -1,36 +1,45 @@
 import {Component, HostBinding, Input, OnInit} from '@angular/core';
 import {MenuService} from '@core/services/menu.service';
-import {animate, animateChild, group, query, sequence, state, style, transition, trigger} from '@angular/animations';
+import {animate, group, query, style, transition, trigger} from '@angular/animations';
 
 @Component({
   selector: 'ebs-submenu',
   templateUrl: './submenu.component.html',
   styleUrls: ['./submenu.component.scss'],
-  animations: [
-    // trigger('routerAnimations', [
-    //   transition('* <=> *', [
-    //     query('@showhideme', animateChild(), {optional: true})
-    //   ]),
-    // ]),
-    trigger('showhideme', [
-      transition('void => *', [
-        style({
-          transform: 'translateY(100%)'
-        }),
-        animate('0.4s', style({
-          transform: 'translateY(0%)'
-        }))
-      ]),
-      transition('* => void', [
-        animate('0.3s', style({
-          transform: 'translateY(100%)'
-        }))
-      ])
-    ])
-  ]
+  // animations: [
+  //   trigger(
+  //     'showHideSubMenu',
+  //     [
+  //       transition(':enter',
+  //         group([
+  //           query('.submenu-container', [
+  //             style({
+  //               opacity: 0,
+  //               transform: 'translateY(100%)'
+  //             }), animate('0.3s',
+  //               style({
+  //                 opacity: 1,
+  //                 transform: 'translateY(0%)'
+  //               }))], {optional: true})
+  //         ])),
+  //       transition(':leave',
+  //         group([
+  //           query('.submenu-container', [
+  //             style({
+  //               opacity: 1,
+  //               transform: 'translateY(0%)'
+  //             }), animate('0.3s',
+  //               style({
+  //                 opacity: 0,
+  //                 transform: 'translateY(100%)'
+  //               }))], {optional: true})
+  //         ]))
+  //     ]
+  //   )
+  // ]
 })
 export class SubmenuComponent implements OnInit {
-  // @HostBinding('@routerAnimations') animateMe: any;
+/*  @HostBinding('@showHideSubMenu') showHideSubMenu: any;*/
 
   @Input() url;
 
