@@ -1,5 +1,5 @@
 export class EnumModel {
   public Name: string;
   public Value: string;
-  public Tag: string;
+  public Id: string;
 }
