@@ -1,4 +1,6 @@
-import { Component } from '@angular/core';
+import {Component} from '@angular/core';
+import {environment} from '../environments/environment';
+import {OAuthService} from 'angular-oauth2-oidc';
 
 @Component({
   selector: 'ebs-root',
@@ -7,4 +9,17 @@ import { Component } from '@angular/core';
 })
 export class AppComponent {
   title = 'ebs';
+
+  public constructor(private oauthService: OAuthService) {
+    this.oauthService.issuer = environment.idpUrl;
+    this.oauthService.clientId = environment.idpUser;
+    this.oauthService.scope = environment.apiScope;
+    this.oauthService.responseType = 'id_token token';
+    this.oauthService.oidc = false;
+    this.oauthService.requireHttps = false;
+    this.oauthService.clearHashAfterLogin = true;
+    this.oauthService.dummyClientSecret = environment.idpPwd;
+    this.oauthService.loadDiscoveryDocument().then(() => {
+    });
+  }
 }
