@@ -1,14 +1,30 @@
 import {NgModule} from '@angular/core';
 import {Routes, RouterModule} from '@angular/router';
+import {AuthGuardService} from '@core/guards/auth-guard.service';
 
-const routes: Routes = [{
-  path: 'anmelden',
-  loadChildren: () => import(`./login/login.module`).then(m => m.LoginModule)
-}, {
-  path: '**',
-  redirectTo: 'anmelden',
-  pathMatch: 'full'
-}];
+const routes: Routes = [
+  {
+    path: 'anmelden',
+    loadChildren: () => import(`./login/login.module`).then(m => m.LoginModule)
+  },
+  {
+    path: 'registrieren',
+    loadChildren: () => import(`./register/register.module`).then(m => m.RegisterModule)
+  },
+  {
+    path: 'benutzerkonto',
+    loadChildren: () => import(`./account/account.module`).then(m => m.AccountModule)
+  },
+  {
+    path: 'ebooks',
+    loadChildren: () => import(`./ebooks/ebooks.module`).then(m => m.EbooksModule)
+  },
+  {
+    path: '**',
+    redirectTo: 'anmelden',
+    pathMatch: 'full'
+  }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes, {
