@@ -1,9 +1,8 @@
 export const environment = {
-  idpUrl: 'https://idp',
-  apiUrl: 'https://api',
+  idpUrl: 'http://127.0.0.1:32839',
+  apiUrl: 'http://127.0.0.1:8080/ebs/api',
   idpUser: 'web',
-  idpPwd: 'D014071B-EEF6-4F49-988B-8E49BC03DF7B',
+    idpPwd: 'D014071B-EEF6-4F49-988B-8E49BC03DF7B',
   apiScope: 'openid profile offline_access webapi',
   production: false,
 };
-
